Move extract_tool_calls to LLM::Session monkey patch
Extract tool call extraction logic from LlmRbFacade into a monkey
patch on LLM::Session in config/initializers/llm_session.rb, so
the session object itself knows how to extract its tool calls.
